// src/app/shared/types/api-types/bire-evolution-variant-dto.interface.ts
export interface BireEvolutionVariantDTO {
  familyCode?: string;
  variantCode?: string;
  evolutionNumber?: string;
  evolutionRevisionNumber?: string;
  evolutionApplicabilityDate?: Date;
  structureType?: string;
  statusStatePerPartner?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}