// src/app/shared/types/api-types/bire-variant-dto.interface.ts
export interface BireVariantDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  variantName?: string;
  variantRevisionNumber?: string;
  variantRevisionDate?: Date;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  familyName?: string;
  familyFunction?: string;
}